separate out createItemTemplate
